pyproject.toml
[tool.poetry]
name = "discord-bot-test"
version = "0.1.0"
description = ""
authors = ["Mizunanari <54758260+Mizunanari@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^1.7.3"
python-dotenv = "^0.20.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
black = "^22.10.0"
pytest = "^7.2.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

# mypyを有効にすると型を指定しないとエラーが表示されます
[tool.mypy]
# check_untyped_defs = true
# disallow_any_generics = true
# disallow_incomplete_defs = true
# disallow_subclassing_any = true
# disallow_untyped_calls = true
# disallow_untyped_defs = true
# no_implicit_optional = true
# no_implicit_reexport = true
# strict_equality = true
# warn_redundant_casts = true
# warn_return_any = true
# warn_unreachable = true
# warn_unused_configs = true
# warn_unused_ignores = true
